+++ services/hipchat/config.go
@@ -12,6 +12,9 @@ type Config struct {
 	Room string `toml:"room"`
 	// Whether all alerts should automatically post to HipChat
 	Global bool `toml:"global"`
+	// Whether all alerts should automatically use stateChangesOnly mode.
+	// Only applies if global is also set.
+	StateChangesOnly bool `toml:"state-changes-only"`
 }
 
 func NewConfig() Config {