+++ services/slack/config.go
@@ -9,6 +9,9 @@ type Config struct {
 	Channel string `toml:"channel"`
 	// Whether all alerts should automatically post to slack
 	Global bool `toml:"global"`
+	// Whether all alerts should automatically use stateChangesOnly mode.
+	// Only applies if global is also set.
+	StateChangesOnly bool `toml:"state-changes-only"`
 }
 
 func NewConfig() Config {